+++ packages/blade-mcp/src/utils.ts
@@ -1,0 +1,35 @@
+import { readFileSync } from 'fs';
+import { join, dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+// Cursor Rules Tokens
+const CURSOR_RULES_VERSION = '0.0.6';
+const CURSOR_RULES_VERSION_STRING = `rules_version: ${CURSOR_RULES_VERSION}`;
+
+const CURSOR_RULES_TEMPLATE_DIRECTORY = join(__dirname, '..', 'cursorRules');
+const BLADE_CURSOR_RULES_FILE_PATH = join(
+  CURSOR_RULES_TEMPLATE_DIRECTORY,
+  'frontend-blade-rules.mdc',
+);
+
+// Blade Template
+const BASE_BLADE_TEMPLATE_DIRECTORY = join(__dirname, '..', 'base-blade-template');
+
+// Knowledgebase
+const KNOWLEDGEBASE_DIRECTORY = join(__dirname, '..', 'knowledgebase');
+
+const hasOutDatedRules = (ruleFilePath: string): boolean => {
+  const ruleFileContent = readFileSync(ruleFilePath, 'utf8');
+  return !ruleFileContent.includes(CURSOR_RULES_VERSION_STRING);
+};
+
+export {
+  BLADE_CURSOR_RULES_FILE_PATH,
+  CURSOR_RULES_VERSION,
+  BASE_BLADE_TEMPLATE_DIRECTORY,
+  KNOWLEDGEBASE_DIRECTORY,
+  hasOutDatedRules,
+};
